+++ flipt-client-browser/src/index.ts
@@ -6,14 +6,16 @@ import {
   BooleanResult,
   EngineOpts,
   EvaluationRequest,
+  IFetcher,
   VariantResult
 } from './models.js';
 
 export class FliptEvaluationClient {
   private engine: Engine;
-  private fetcher: () => Promise<Response>;
+  private fetcher: IFetcher;
+  private etag?: string;
 
-  private constructor(engine: Engine, fetcher: () => Promise<Response>) {
+  private constructor(engine: Engine, fetcher: IFetcher) {
     this.engine = engine;
     this.fetcher = fetcher;
   }
@@ -44,7 +46,7 @@ export class FliptEvaluationClient {
 
     const headers = new Headers();
     headers.append('Accept', 'application/json');
-    headers.append('Content-Type', 'application/json');
+    headers.append('x-flipt-accept-server-version', '1.47.0');
 
     if (engine_opts.authentication) {
       if ('client_token' in engine_opts.authentication) {
@@ -63,11 +65,22 @@ export class FliptEvaluationClient {
     let fetcher = engine_opts.fetcher;
 
     if (!fetcher) {
-      fetcher = async () => {
+      fetcher = async (opts?: { etag?: string }) => {
+        if (opts && opts.etag) {
+          headers.append('If-None-Match', opts.etag);
+        }
+
         const resp = await fetch(url, {
           method: 'GET',
           headers
         });
+
+        // check for 304 status code
+        if (resp.status === 304) {
+          return resp;
+        }
+
+        // ok only checks for range 200-299
         if (!resp.ok) {
           throw new Error('Failed to fetch data');
         }
@@ -76,7 +89,12 @@ export class FliptEvaluationClient {
       };
     }
 
+    // should be no etag on first fetch
     const resp = await fetcher();
+    if (!resp) {
+      throw new Error('Failed to fetch data');
+    }
+
     const data = await resp.json();
 
     const engine = new Engine(namespace, data);
@@ -88,9 +106,18 @@ export class FliptEvaluationClient {
    * @returns void
    */
   public async refresh() {
-    const resp = await this.fetcher();
-    const data = await resp.json();
+    const opts = { etag: this.etag };
+    const resp = await this.fetcher(opts);
 
+    if (resp.status === 304) {
+      let etag = resp.headers.get('etag');
+      if (etag) {
+        this.etag = etag;
+      }
+      return;
+    }
+
+    const data = await resp.json();
     this.engine.snapshot(data);
   }