+++ flipt-evaluation/src/lib.rs
@@ -135,6 +135,18 @@ pub fn variant_evaluation(
         ..Default::default()
     };
 
+    if flag.default_variant.is_some() {
+        variant_evaluation_response.reason = common::EvaluationReason::Default;
+
+        let default_variant = flag.default_variant.as_ref().unwrap();
+        variant_evaluation_response
+            .variant_key
+            .clone_from(&default_variant.key);
+        variant_evaluation_response
+            .variant_attachment
+            .clone_from(&default_variant.attachment);
+    }
+
     if !flag.enabled {
         variant_evaluation_response.reason = common::EvaluationReason::FlagDisabled;
         variant_evaluation_response.request_duration_millis = get_duration_millis(now)?;
@@ -151,6 +163,12 @@ pub fn variant_evaluation(
             )));
         }
     };
+
+    // if no rules and flag is enabled, return default variant
+    if evaluation_rules.is_empty() {
+        variant_evaluation_response.request_duration_millis = get_duration_millis(now)?;
+        return Ok(variant_evaluation_response);
+    }
 
     for rule in evaluation_rules {
         if rule.rank < last_rank {
@@ -224,6 +242,7 @@ pub fn variant_evaluation(
         }
 
         // no distributions for the rule
+        // match is true here because it did match the segment/rule
         if valid_distributions.is_empty() {
             variant_evaluation_response.r#match = true;
             variant_evaluation_response.reason = common::EvaluationReason::Match;
@@ -242,6 +261,7 @@ pub fn variant_evaluation(
             Err(idx) => idx,
         };
 
+        // if index is outside of our existing buckets then it does not match any distribution
         if index == valid_distributions.len() {
             variant_evaluation_response.r#match = false;
             variant_evaluation_response.request_duration_millis = get_duration_millis(now)?;
@@ -1006,6 +1026,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: true,
                 r#type: common::FlagType::Variant,
+                default_variant: None,
             })
         });
 
@@ -1061,6 +1082,89 @@ mod tests {
         assert_eq!(v.segment_keys, vec![String::from("segment1")]);
     }
 
+    #[test]
+    fn test_evaluator_flag_disabled() {
+        let mut mock_store = MockStore::new();
+
+        mock_store.expect_get_flag().returning(|_, _| {
+            Some(flipt::Flag {
+                key: String::from("foo"),
+                enabled: false,
+                r#type: common::FlagType::Variant,
+                default_variant: None,
+            })
+        });
+
+        let mut context: HashMap<String, String> = HashMap::new();
+        context.insert(String::from("bar"), String::from("baz"));
+        context.insert(String::from("foo"), String::from("bar"));
+
+        let variant = variant_evaluation(
+            &mock_store,
+            "default",
+            &EvaluationRequest {
+                flag_key: String::from("foo"),
+                entity_id: String::from("entity"),
+                context,
+            },
+        );
+
+        assert!(variant.is_ok());
+
+        let v = variant.unwrap();
+
+        assert_eq!(v.flag_key, String::from("foo"));
+        assert!(!v.r#match);
+        assert_eq!(v.reason, common::EvaluationReason::FlagDisabled);
+        assert!(v.variant_key.is_empty());
+        assert!(v.variant_attachment.is_empty());
+    }
+
+    #[test]
+    fn test_evaluator_flag_disabled_default_variant() {
+        let mut mock_store = MockStore::new();
+
+        mock_store.expect_get_flag().returning(|_, _| {
+            Some(flipt::Flag {
+                key: String::from("foo"),
+                enabled: false,
+                r#type: common::FlagType::Variant,
+                default_variant: Some(flipt::Variant {
+                    id: String::from("1"),
+                    key: String::from("default"),
+                    attachment: serde_json::json!({"key": "value"}).to_string(),
+                }),
+            })
+        });
+
+        let mut context: HashMap<String, String> = HashMap::new();
+        context.insert(String::from("bar"), String::from("baz"));
+        context.insert(String::from("foo"), String::from("bar"));
+
+        let variant = variant_evaluation(
+            &mock_store,
+            "default",
+            &EvaluationRequest {
+                flag_key: String::from("foo"),
+                entity_id: String::from("entity"),
+                context,
+            },
+        );
+
+        assert!(variant.is_ok());
+
+        let v = variant.unwrap();
+
+        assert_eq!(v.flag_key, String::from("foo"));
+        assert!(!v.r#match);
+        assert_eq!(v.reason, common::EvaluationReason::FlagDisabled);
+        assert_eq!(v.variant_key, String::from("default"));
+        assert_eq!(
+            v.variant_attachment,
+            serde_json::json!({"key": "value"}).to_string()
+        );
+    }
+
     // Segment Match Type ALL
     #[test]
     fn test_evaluator_match_all_no_variants_no_distributions() {
@@ -1071,6 +1175,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: true,
                 r#type: common::FlagType::Variant,
+                default_variant: None,
             })
         });
 
@@ -1137,6 +1242,90 @@ mod tests {
     }
 
     #[test]
+    fn test_evaluator_match_all_no_distributions_default_variant() {
+        let mut mock_store = MockStore::new();
+
+        mock_store.expect_get_flag().returning(|_, _| {
+            Some(flipt::Flag {
+                key: String::from("foo"),
+                enabled: true,
+                r#type: common::FlagType::Variant,
+                default_variant: Some(flipt::Variant {
+                    id: String::from("1"),
+                    key: String::from("default"),
+                    attachment: serde_json::json!({"key": "value"}).to_string(),
+                }),
+            })
+        });
+
+        let mut segments: HashMap<String, flipt::EvaluationSegment> = HashMap::new();
+        segments.insert(
+            String::from("segment1"),
+            flipt::EvaluationSegment {
+                segment_key: String::from("segment1"),
+                match_type: common::SegmentMatchType::All,
+                constraints: vec![
+                    flipt::EvaluationConstraint {
+                        r#type: common::ConstraintComparisonType::String,
+                        property: String::from("bar"),
+                        operator: String::from("eq"),
+                        value: String::from("baz"),
+                    },
+                    flipt::EvaluationConstraint {
+                        r#type: common::ConstraintComparisonType::String,
+                        property: String::from("foo"),
+                        operator: String::from("eq"),
+                        value: String::from("bar"),
+                    },
+                ],
+            },
+        );
+
+        mock_store
+            .expect_get_evaluation_rules()
+            .returning(move |_, _| {
+                Some(vec![flipt::EvaluationRule {
+                    id: String::from("1"),
+                    flag_key: String::from("foo"),
+                    segments: segments.clone(),
+                    rank: 1,
+                    segment_operator: common::SegmentOperator::Or,
+                }])
+            });
+
+        mock_store
+            .expect_get_evaluation_distributions()
+            .returning(|_, _| Some(vec![]));
+
+        let mut context: HashMap<String, String> = HashMap::new();
+        context.insert(String::from("bar"), String::from("baz"));
+        context.insert(String::from("foo"), String::from("bar"));
+
+        let variant = variant_evaluation(
+            &mock_store,
+            "default",
+            &EvaluationRequest {
+                flag_key: String::from("foo"),
+                entity_id: String::from("entity"),
+                context,
+            },
+        );
+        assert!(variant.is_ok());
+
+        let v = variant.unwrap();
+
+        assert_eq!(v.flag_key, String::from("foo"));
+        assert!(v.r#match);
+        assert_eq!(v.reason, common::EvaluationReason::Match);
+        assert_eq!(v.segment_keys, vec![String::from("segment1")]);
+        assert_eq!(v.variant_key, String::from("default"));
+        assert_eq!(
+            v.variant_attachment,
+            serde_json::json!({"key": "value"}).to_string()
+        );
+    }
+
+    #[test]
     fn test_evaluator_match_all_multiple_segments() {
         let mut mock_store = MockStore::new();
 
@@ -1145,6 +1334,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: true,
                 r#type: common::FlagType::Variant,
+                default_variant: None,
             })
         });
 
@@ -1256,6 +1446,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: true,
                 r#type: common::FlagType::Variant,
+                default_variant: None,
             })
         });
 
@@ -1351,6 +1542,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: true,
                 r#type: common::FlagType::Variant,
+                default_variant: None,
             })
         });
 
@@ -1456,6 +1648,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: true,
                 r#type: common::FlagType::Variant,
+                default_variant: None,
             })
         });
 
@@ -1567,6 +1760,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: true,
                 r#type: common::FlagType::Variant,
+                default_variant: None,
             })
         });
 
@@ -1677,6 +1871,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: true,
                 r#type: common::FlagType::Variant,
+                default_variant: None,
             })
         });
 
@@ -1774,6 +1969,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: true,
                 r#type: common::FlagType::Variant,
+                default_variant: None,
             })
         });
 
@@ -1840,6 +2036,90 @@ mod tests {
     }
 
     #[test]
+    fn test_evaluator_match_any_no_distributions_default_variant() {
+        let mut mock_store = MockStore::new();
+
+        mock_store.expect_get_flag().returning(|_, _| {
+            Some(flipt::Flag {
+                key: String::from("foo"),
+                enabled: true,
+                r#type: common::FlagType::Variant,
+                default_variant: Some(flipt::Variant {
+                    id: String::from("1"),
+                    key: String::from("default"),
+                    attachment: serde_json::json!({"key": "value"}).to_string(),
+                }),
+            })
+        });
+
+        let mut segments: HashMap<String, flipt::EvaluationSegment> = HashMap::new();
+        segments.insert(
+            String::from("segment1"),
+            flipt::EvaluationSegment {
+                segment_key: String::from("segment1"),
+                match_type: common::SegmentMatchType::Any,
+                constraints: vec![
+                    flipt::EvaluationConstraint {
+                        r#type: common::ConstraintComparisonType::String,
+                        property: String::from("bar"),
+                        operator: String::from("eq"),
+                        value: String::from("baz"),
+                    },
+                    flipt::EvaluationConstraint {
+                        r#type: common::ConstraintComparisonType::String,
+                        property: String::from("foo"),
+                        operator: String::from("eq"),
+                        value: String::from("bar"),
+                    },
+                ],
+            },
+        );
+
+        mock_store
+            .expect_get_evaluation_rules()
+            .returning(move |_, _| {
+                Some(vec![flipt::EvaluationRule {
+                    id: String::from("1"),
+                    flag_key: String::from("foo"),
+                    segments: segments.clone(),
+                    rank: 1,
+                    segment_operator: common::SegmentOperator::Or,
+                }])
+            });
+
+        mock_store
+            .expect_get_evaluation_distributions()
+            .returning(|_, _| Some(vec![]));
+
+        let mut context: HashMap<String, String> = HashMap::new();
+        context.insert(String::from("bar"), String::from("baz"));
+
+        let variant = variant_evaluation(
+            &mock_store,
+            "default",
+            &EvaluationRequest {
+                flag_key: String::from("foo"),
+                entity_id: String::from("entity"),
+                context,
+            },
+        );
+
+        assert!(variant.is_ok());
+
+        let v = variant.unwrap();
+
+        assert_eq!(v.flag_key, String::from("foo"));
+        assert!(v.r#match);
+        assert_eq!(v.reason, common::EvaluationReason::Match);
+        assert_eq!(v.segment_keys, vec![String::from("segment1")]);
+        assert_eq!(v.variant_key, String::from("default"));
+        assert_eq!(
+            v.variant_attachment,
+            serde_json::json!({"key": "value"}).to_string()
+        );
+    }
+
+    #[test]
     fn test_evaluator_match_any_multiple_segments() {
         let mut mock_store = MockStore::new();
 
@@ -1848,6 +2128,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: true,
                 r#type: common::FlagType::Variant,
+                default_variant: None,
             })
         });
 
@@ -1957,6 +2238,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: true,
                 r#type: common::FlagType::Variant,
+                default_variant: None,
             })
         });
 
@@ -2051,6 +2333,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: true,
                 r#type: common::FlagType::Variant,
+                default_variant: None,
             })
         });
 
@@ -2147,6 +2430,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: true,
                 r#type: common::FlagType::Variant,
+                default_variant: None,
             })
         });
 
@@ -2257,6 +2541,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: true,
                 r#type: common::FlagType::Variant,
+                default_variant: None,
             })
         });
 
@@ -2366,6 +2651,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: true,
                 r#type: common::FlagType::Variant,
+                default_variant: None,
             })
         });
 
@@ -2485,6 +2771,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: true,
                 r#type: common::FlagType::Variant,
+                default_variant: None,
             })
         });
 
@@ -2567,6 +2854,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: true,
                 r#type: common::FlagType::Variant,
+                default_variant: None,
             })
         });
 
@@ -2673,6 +2961,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: false,
                 r#type: common::FlagType::Boolean,
+                default_variant: None,
             })
         });
 
@@ -2734,6 +3023,7 @@ mod tests {
                 key: String::from("foo"),
                 enabled: false,
                 r#type: common::FlagType::Boolean,
+                default_variant: None,
             })
         });